+++ src/BugsnagUnity/Native/Cocoa/NativeCode.cs
@@ -74,5 +74,11 @@ namespace BugsnagUnity
 
     [DllImport(Import)]
     internal static extern void bugsnag_retrieveBreadcrumbs(IntPtr breadcrumbs, IntPtr instance, BreadcrumbInformation visitor);
+
+    [DllImport(Import)]
+    internal static extern void bugsnag_setUser(string id, string name, string email);
+
+    [DllImport(Import)]
+    internal static extern void bugsnag_registerSession(string id, long startedAt, int unhandledCount, int handledCount);
   }
 }